backend/requirements.txt
# Web framework
fastapi
uvicorn
python-dotenv

# Supabase
supabase

# LlamaIndex
llama-index
llama-index-vector-stores-supabase

# LangChain / LangGraph
langchain
langgraph

# Observability & evaluation
langfuse
ragas

# Testing
pytest>=7.0
pytest-cov>=4.0

# AI providers
openai
anthropic
google-generativeai

# HTTP & data
requests
beautifulsoup4
feedparser
httpx
pandas
pandas-ta>=0.4.67b0
